streamlit-requirements.txt
streamlit==1.32.0
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pgvector==0.2.4
